requirements.txt
Django>=5.0,<5.2
djangorestframework>=3.15,<3.17
djangorestframework-simplejwt>=5.3,<5.6
drf-spectacular>=0.27,<0.30
django-filter>=24.0,<26.0
django-cors-headers>=4.3,<5.0
python-decouple>=3.8,<4.0
psycopg2-binary>=2.9,<3.0
Pillow>=10.0,<12.0

# Dev
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
